# Optimizes various components