--- a/src/site/_includes/affordable-housing-item.liquid
+++ b/src/site/_includes/affordable-housing-item.liquid
@@ -27,11 +27,9 @@ https://maps.google.com/maps?q={{ address.address | url_encode}}+{{ address.city
   {%- if address.address and address.address != '' -%}
     <p class="contact">
     {{ address.address }}{%- if address.city and address.city != '' -%}, {{ address.city }}{%- endif -%}
-      {%- unless isInfoWindow -%}
-        <span class="map_link_container noprint nowrap">
-          (<button type="button" class="map_link link hidden" aria-label="Highlight {{ address.aptName }} on map">Show on map</button><span class="ext_map_link"><a href="{{mapUrl}}" target="_blank" rel="noopener" aria-label="View {{ address.aptName }} on Google Maps">map</a><sup>&#8599;</sup></span>)
-        </span>
-      {%- endunless -%}
+      <span class="map_link_container noprint nowrap">
+        (<button type="button" class="map_link link hidden" aria-label="Highlight {{ address.aptName }} on map">Show on map</button><span class="ext_map_link"><a href="{{mapUrl}}" target="_blank" rel="noopener" aria-label="View {{ address.aptName }} on Google Maps">map</a><sup>&#8599;</sup></span>)
+      </span>
     </p>
   {%- endif -%}
   {%- if address.email and address.email != '' -%}
@@ -48,10 +46,5 @@ https://maps.google.com/maps?q={{ address.address | url_encode}}+{{ address.city
   {%- if address.website and address.website != '' -%}
     <p class="contact">
       <a href="{{ address.website }}" target="_blank" rel="noopener" aria-label="Property website for {{ address.aptName }}">View property website</a><sup>&#8599;</sup>
-    </p>
-  {%- endif -%}
-  {% if isInfoWindow and address.address and address.address != '' %}
-    <p class="map_infowindow_footer">
-      <a href="{{mapUrl}}" target="_blank" rel="noopener" aria-label="View {{address.aptName}} on Google Maps">View on Google Maps</a><sup>&#8599;</sup>
     </p>
   {%- endif -%}