--- a/src/site/_includes/fonts.scss
+++ b/src/site/_includes/fonts.scss
@@ -5,4 +5,5 @@
   unicode-range: U+20-7E;
   font-weight: 100 1000;
   font-stretch: 25% 151%;
+  font-display: fallback;
 }