--- a/src/site/_includes/header.liquid
+++ b/src/site/_includes/header.liquid
@@ -1,6 +1,6 @@
   
   <a href="/" class="nav-home" title="Home">
-    <img src="/images/ueo-logo.png" class="logo-home" alt="The United Effort Organization logo">
+    <img src="/images/ueo-logo.png" width="200" height="61" class="logo-home" alt="The United Effort Organization logo">
   </a>
   <nav>
     <ul aria-label="Navigation Links" class="navigation">