--- a/src/site/_includes/layouts/base.liquid
+++ b/src/site/_includes/layouts/base.liquid
@@ -6,6 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>{% if title and title != '' %}{{title}} - {% endif %}The United Effort Organization</title>
   <link rel="stylesheet" href="/styles.css">
+  <link rel="preload" href="/fonts/Inter-Variable-subset.woff2" as="font" type="font/woff2" crossorigin>
   <link rel="apple-touch-icon" sizes="180x180" href="/favicon/apple-touch-icon.png">
   <link rel="icon" type="image/png" sizes="32x32" href="/favicon/favicon-32x32.png">
   <link rel="icon" type="image/png" sizes="16x16" href="/favicon/favicon-16x16.png">