--- a/src/site/_includes/type.scss
+++ b/src/site/_includes/type.scss
@@ -226,6 +226,9 @@
       margin: 0;
       padding: 0;
     }
+    p.contact {
+      line-height: 1.8em;
+    }
     h2 {
       display: inline-block;
       font-size: 1.3em;