--- a/src/site/_includes/type.scss
+++ b/src/site/_includes/type.scss
@@ -258,6 +258,7 @@
     .map_infowindow_footer {
       padding-top: 1.5em;
       font-size: 0.8em;
+      display: block;
     }
   }