--- a/src/site/serverless/affordable-housing.liquid
+++ b/src/site/serverless/affordable-housing.liquid
@@ -222,7 +222,7 @@ Narrow your search with the filter options below or <a href="#results">scroll</a
       <ul class="results_list" aria-label="Results List">
         {% for address in sortedHousingdb %}
           <li id="property-{{address.id}}">
-            {% render "affordable-housing-item.liquid", address: address, isInfoWindow: false %}
+            {% render "affordable-housing-item.liquid", address: address %}
           </li>
         {% endfor %}
         {% if num_results == 0 %}
@@ -253,12 +253,7 @@ Narrow your search with the filter options below or <a href="#results">scroll</a
   {% for address in sortedHousingdb %}
     {% assign coords = address | getValidatedLocCoords %}
     {% if coords %}
-      {% capture rawContent %}
-        <div id="infowindow-{{address.id}}" class="map_infowindow_content">
-          {% render "affordable-housing-item.liquid", address: address, isInfoWindow: true %}
-        </div> 
-      {% endcapture %}
-      { lat: {{coords[0]}}, lng: {{coords[1]}}, content: '{{ rawContent | replace: "\n", "" | replace: "'", "&apos;"}}', id: "{{address.id}}" },
+      {lat:{{coords[0]}},lng:{{coords[1]}},id:"{{address.id}}"},
     {% endif %}
   {% endfor %}
   ];