--- a/src/site/serverless/affordable-housing.liquid
+++ b/src/site/serverless/affordable-housing.liquid
@@ -6,6 +6,7 @@ permalink:
   serverless: "/housing/affordable-housing/filter"
 pageClass: "page-housing"
 pageHead: '<script src="/js/accessibility.js" defer></script>
+  <script src="/js/transit.js" defer></script>
   <script src="/js/affordable-housing-map.js" defer></script>
   <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyA17udd4ePocbxKrOkHiMhSUHlvf3daZMs&callback=initMap&v=weekly" defer></script>'
 ---
@@ -255,12 +256,6 @@ Narrow your search with the filter options below or <a href="#results">scroll</a
     {% if coords %}
       {lat:{{coords[0]}},lng:{{coords[1]}},id:"{{address.id}}"},
     {% endif %}
-  {% endfor %}
-  ];
-
-  const transitStops = [
-  {% for stop in transitData %}
-    { lat: {{stop.lat}}, lng: {{stop.lng}}, name: "{{stop.name}}", operator: "{{stop.operator}}" },
   {% endfor %}
   ];
 </script>