… src/assets/images/bus_small_2x.png
… src/assets/images/measle_2x.png
… src/assets/images/ueo-logo.png
… src/site/serverless/transit.11tydata.js