Overall result:

- Lighthouse score (emulated Moto G4, slow 4G throttling): 43 --> 65
  - First contentful paint: 1.8s --> 1.3s
  - Speed index: 4.5s --> 2.0s
  - Time to interactive: 10.1s --> 7.2s
  - Total blocking time: 7070ms -> 4170ms
- affordable housing index.html size: 1.4MB --> 550KB
- affordable housing index.html DOM elements: 9,682 --> 8,416